+++ packages/expo-video/src/VideoView.web.tsx
@@ -4,6 +4,13 @@ import { StyleSheet } from 'react-native';
 import VideoPlayer, { getSourceUri } from './VideoPlayer.web';
 import type { VideoViewProps } from './VideoView.types';
 
+type FullscreenChangeListeners = {
+  default: () => void;
+  safariEnter: () => void;
+  safariExit: () => void;
+  msListener: () => void;
+};
+
 function createAudioContext(): AudioContext | null {
   return typeof window !== 'undefined' ? new window.AudioContext() : null;
 }
@@ -32,7 +39,7 @@ export const VideoView = forwardRef((props: { player?: VideoPlayer } & VideoView
   const videoRef = useRef<null | HTMLVideoElement>(null);
   const mediaNodeRef = useRef<null | MediaElementAudioSourceNode>(null);
   const hasToSetupAudioContext = useRef(false);
-  const fullscreenChangeListener = useRef<null | (() => void)>(null);
+  const fullscreenChangeListeners = useRef<null | FullscreenChangeListeners>(null);
   const isWaitingForFirstFrame = useRef(false);
 
   /**
@@ -55,10 +62,23 @@ export const VideoView = forwardRef((props: { player?: VideoPlayer } & VideoView
 
   useImperativeHandle(ref, () => ({
     enterFullscreen: async () => {
-      if (!props.allowsFullscreen) {
+      if (!props.allowsFullscreen || !videoRef.current) {
         return;
       }
-      await videoRef.current?.requestFullscreen();
+      // Cast the video to any to avoid ts errors. Methods such as webkitRequestFullscreen,
+      // webkitEnterFullScreen, msRequestFullscreen are not typed even though they exist.
+      const video = videoRef.current as any;
+
+      if (video.requestFullscreen) {
+        await video.requestFullscreen();
+      } else if (video.webkitRequestFullscreen) {
+        // @ts-ignore webkitRequestFullscreen can exist on Apple devices
+        await video.webkitRequestFullscreen();
+      } else if (video.webkitEnterFullScreen) {
+        await video.webkitEnterFullScreen();
+      } else if (video.msRequestFullscreen) {
+        await video.msRequestFullscreen();
+      }
     },
     exitFullscreen: async () => {
       await document.exitFullscreen();
@@ -168,15 +188,43 @@ export const VideoView = forwardRef((props: { player?: VideoPlayer } & VideoView
   }
 
   function setupFullscreenListener() {
-    fullscreenChangeListener.current = fullscreenListener;
-    videoRef.current?.addEventListener('fullscreenchange', fullscreenChangeListener.current);
+    cleanupFullscreenListener();
+    const video = videoRef.current;
+    if (!video) return;
+    const fullscreenListeners: FullscreenChangeListeners = {
+      default: fullscreenListener,
+      safariEnter: () => props.onFullscreenEnter?.(),
+      safariExit: () => props.onFullscreenExit?.(),
+      msListener: fullscreenListener,
+    };
+
+    fullscreenChangeListeners.current = fullscreenListeners;
+
+    // Standard Fullscreen API
+    video.addEventListener('fullscreenchange', fullscreenListeners.default);
+
+    // Safari (webkit)
+    video.addEventListener('webkitbeginfullscreen', fullscreenListeners.safariEnter);
+    video.addEventListener('webkitendfullscreen', fullscreenListeners.safariExit);
+
+    // IE11 (ms)
+    document.addEventListener('MSFullscreenChange', fullscreenListeners.msListener);
   }
 
   function cleanupFullscreenListener() {
-    if (fullscreenChangeListener.current) {
-      videoRef.current?.removeEventListener('fullscreenchange', fullscreenChangeListener.current);
-      fullscreenChangeListener.current = null;
-    }
+    const video = videoRef.current;
+    if (!video || !fullscreenChangeListeners.current) return;
+
+    video.removeEventListener('fullscreenchange', fullscreenChangeListeners.current.default);
+    video.removeEventListener(
+      'webkitbeginfullscreen',
+      fullscreenChangeListeners.current?.safariEnter
+    );
+    video.removeEventListener('webkitendfullscreen', fullscreenChangeListeners.current.safariExit);
+    document.removeEventListener(
+      'MSFullscreenChange',
+      fullscreenChangeListeners.current.msListener
+    );
   }
 
   useEffect(() => {